Squashed commit of the following:
commit 23c0aa2c084efe5e71274f52ebfeb17c64263885
Author: Taras Kushnir <kushnirTV@gmail.com>
Date:   Tue Jul 5 16:36:02 2016 +0300

    Fixes for tests

commit e3f598b75ccd170fe55d41b6e8ee0ab9a451eeaf
Author: Taras Kushnir <kushnirTV@gmail.com>
Date:   Tue Jul 5 16:09:26 2016 +0300

    Uncommented core tests

commit e4cc74e99da981fd324def3fbf5d0867d2d75a7a
Author: Taras Kushnir <kushnirTV@gmail.com>
Date:   Tue Jul 5 16:07:52 2016 +0300

    Integration tests warnings fixes

commit f51885e491c1cf6c8fef20f3e6aa2c98a22b9ff2
Author: Taras Kushnir <kushnirTV@gmail.com>
Date:   Tue Jul 5 16:04:11 2016 +0300

    Core tests fixed

commit f1029b53a222fb4a85a409527a48781b08d7cdcb
Author: Taras Kushnir <kushnirTV@gmail.com>
Date:   Tue Jul 5 15:24:56 2016 +0300

    Refactoring for smart pointers

+++ src/xpiks-qt/Common/iartworkssource.h
@@ -30,7 +30,7 @@ namespace Common {
         virtual ~IArtworksSource() {}
 
         virtual IBasicArtwork *getBasicArtwork(int index) const = 0;
-        virtual int getArtworksCount() const = 0;
+        virtual size_t getArtworksCount() const = 0;
     };
 }
